Fix stitched decrypt
Observe that the stitched decrypt code isn't normally generated, and
didn't take the refactoring well due to the excessive use of mutable
globals.

Further, there is a "bug" in the conditional disabling. The 'if
$stitched_decrypt' moved the global @aes256_dec table to local scope and
broke its redefinitions, which ended up generating the wrong code for
AVX when setting "$stitched_decrypt = 1". Move the declaration of this
variable to global scope.

Of course in reality, the decryption code is dead and should be deleted.